pyproject.toml
[project]
name = "lrcbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apropos-ai>=0.1.10",
    "sympy>=1.13.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
